# Add support for caching CMD BO objects used in xclCopyBO() for both P…