--- a/src/runtime_src/core/edge/user/shim.cpp
+++ b/src/runtime_src/core/edge/user/shim.cpp
@@ -2,7 +2,7 @@
  * Copyright (C) 2016-2019 Xilinx, Inc
  * Author(s): Hem C. Neema
  *          : Min Ma
- * ZNYQ HAL Driver layered on top of ZYNQ kernel driver
+ * ZNYQ XRT Library layered on top of ZYNQ zocl kernel driver
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may
  * not use this file except in compliance with the License. A copy of the
@@ -35,7 +35,8 @@
 #include "core/common/message.h"
 #include "core/common/scheduler.h"
 #include "core/common/xclbin_parser.h"
-//#include "xclbin.h"
+#include "core/common/bo_cache.h"
+#include "core/common/config_reader.h"
 #include <assert.h>
 
 #define GB(x)   ((size_t) (x) << 30)
@@ -117,12 +118,14 @@ ZYNQShim::ZYNQShim(unsigned index, const char *logfileName, xclVerbosityLevel ve
     mLogStream << "FUNCTION, THREAD ID, ARG..." << std::endl;
     mLogStream << __func__ << ", " << std::this_thread::get_id() << std::endl;
   }
+  mCmdBOCache = new xrt_core::bo_cache(this, xrt_core::config::get_cmdbo_cache());
 }
 
 #ifndef __HWEM__
 ZYNQShim::~ZYNQShim()
 {
-  if (profiling != nullptr) delete profiling;
+  delete mCmdBOCache;
+  delete profiling;
   //TODO
   if (mKernelFD > 0) {
     close(mKernelFD);
@@ -378,6 +381,38 @@ int ZYNQShim::xclSyncBO(unsigned int boHandle, xclBOSyncDirection dir, size_t si
   return ioctl(mKernelFD, DRM_IOCTL_ZOCL_SYNC_BO, &syncInfo);
 }
 
+int ZYNQShim::xclCopyBO(unsigned int dst_boHandle, unsigned int src_boHandle, size_t size,
+                        size_t dst_offset, size_t src_offset)
+{
+  int ret = -EOPNOTSUPP;
+#ifdef __aarch64__
+    std::pair<unsigned int, ert_start_copybo_cmd *const> bo = mCmdBOCache->alloc<ert_start_copybo_cmd>();
+    ert_fill_copybo_cmd(bo.second, src_boHandle, dst_boHandle,
+                        src_offset, dst_offset, size);
+
+    int ret = xclExecBuf(bo.first);
+    if (ret == 0)
+        while (xclExecWait(1000) == 0);   if (ret) {
+        mCmdBOCache->release(bo);
+        return ret;
+    }
+
+    do {
+        ret = xclExecWait(1000);
+        if (ret == -1)
+            break;
+    }
+    while (bo.second->state < ERT_CMD_STATE_COMPLETED);
+
+    ret = (ret == -1) ? -errno : 0;
+    if (!ret && (bo.second->state != ERT_CMD_STATE_COMPLETED))
+        ret = -EINVAL;
+    mCmdBOCache->release<ert_start_copybo_cmd>(bo);
+#endif
+  return ret;
+}
+
+
 #ifndef __HWEM__
 int ZYNQShim::xclLoadXclBin(const xclBin *buffer)
 {
@@ -530,14 +565,14 @@ uint ZYNQShim::xclGetNumLiveProcesses()
 int ZYNQShim::xclGetSysfsPath(const char* subdev, const char* entry, char* sysfsPath, size_t size)
 {
   // Until we have a programmatic way to determine what this directory
-  //  is on Zynq platforms, this is hard-coded so we can test out 
+  //  is on Zynq platforms, this is hard-coded so we can test out
   //  debug and profile features.
   std::string path = "/sys/devices/platform/amba/amba:zyxclmm_drm/";
   path += entry ;
 
   if (path.length() >= size) return -1 ;
 
-  // Since path.length() < size, we are sure to copy over the null 
+  // Since path.length() < size, we are sure to copy over the null
   //  terminating byte.
   strncpy(sysfsPath, path.c_str(), size) ;
   return 0 ;
@@ -586,7 +621,7 @@ int ZYNQShim::xclSKReport(uint32_t cu_idx, xrt_scu_state state)
   }
 
   scmd.cu_idx = cu_idx;
-  
+
   ret = ioctl(mKernelFD, DRM_IOCTL_ZOCL_SK_REPORT, &scmd);
 
   return ret;
@@ -835,7 +870,7 @@ uint xclGetNumLiveProcesses(xclDeviceHandle handle)
   return drv->xclGetNumLiveProcesses();
 }
 
-int xclGetSysfsPath(xclDeviceHandle handle, const char* subdev, 
+int xclGetSysfsPath(xclDeviceHandle handle, const char* subdev,
 		    const char* entry, char* sysfsPath, size_t size)
 {
   ZYNQ::ZYNQShim *drv = ZYNQ::ZYNQShim::handleCheck(handle);