--- a/src/runtime_src/core/pcie/linux/shim.cpp
+++ b/src/runtime_src/core/pcie/linux/shim.cpp
@@ -440,7 +440,6 @@ int shim::xclCopyBO(unsigned int dst_boHandle,
                         src_offset, dst_offset, size);
 
     int ret = xclExecBuf(bo.first);
-
     if (ret) {
         mCmdBOCache->release(bo);
         return ret;