…CIe and edge platforms.

Update xbutil to use xclCopyBO() for M2M test. The size of the cache is determined by Runtime.cmdbo_cache
key in ert.ini and defaults to 4.

xclCopyBO() with BO Cache improves the throughput by hiding the CMD BO management
overhead. See the xbutil m2mtest metrics below for 256 KB buffers:
After change
------------
dx4300:~/git/XRT/build>/home/sonals/git/XRT/build/Release/opt/xilinx/xrt/bin/xbutil m2mtest
INFO: Found total 1 card(s), 1 are usable
bank0 -> bank1 M2M bandwidth: 833.333 MB/s
bank0 -> bank2 M2M bandwidth: 3086.42 MB/s
bank0 -> bank3 M2M bandwidth: 3623.19 MB/s
bank1 -> bank2 M2M bandwidth: 3787.88 MB/s
bank1 -> bank3 M2M bandwidth: 3731.34 MB/s
bank2 -> bank3 M2M bandwidth: 3787.88 MB/s
INFO: xbutil m2mtest succeeded.
dx4300:~/git/XRT/build>

Before change
-------------
dx4300:~/git/XRT/build>/home/sonals/git/XRT/build/Release/opt/xilinx/xrt/bin/xbutil m2mtest
INFO: Found total 1 card(s), 1 are usable
bank0 -> bank1 M2M bandwidth: 868.056 MB/s
bank0 -> bank2 M2M bandwidth: 1201.92 MB/s
bank0 -> bank3 M2M bandwidth: 1552.8 MB/s
bank1 -> bank2 M2M bandwidth: 1724.14 MB/s
bank1 -> bank3 M2M bandwidth: 1436.78 MB/s
bank2 -> bank3 M2M bandwidth: 696.379 MB/s
INFO: xbutil m2mtest succeeded.
dx4300:~/git/XRT/build>