@maxzhen 2019-07-14
On what board did you obtain these m2m perf numbers? On vcu1525 and U200, they are 12GB/s and very stable....