@sonals 2019-07-15
> On what board did you obtain these m2m perf numbers? On vcu1525 and U200, they are 12GB/s and very stable....

These for 256 KB buffers where overhead is more visible :-)
For default 256 MB buffers I do get 12 GB/s as well.